не работает

+++ VoiceTexterBot/Configuration/AppSettings.cs
@@ -32,6 +32,10 @@ namespace VoiceTexterBot.Configuration
         /// Формат аудио при загрузке
         /// </summary>
         public object OutputAudioFormat { get; set; }
-        public int InputAudioBitrate { get; set; }
+
+        /// <summary>
+        /// Битрейт аудио при загрузке
+        /// </summary>
+        public float InputAudioBitrate { get; set; }
     }
 }
